Multicoin (#495)
* Abstract TextRecord component to a reusable component

* WIP - address

* Add queries and mutations

* AddRecord for multicoin

* WIP, remove ETH

* Upgrade package.json

* Fix add content test

* fix integration test

* revert
