+++ package.json
@@ -12,7 +12,7 @@
     "@ensdomains/ethregistrar": "^1.2.2",
     "@ensdomains/react-ens-address": "^0.0.16",
     "@ensdomains/resolver": "^0.1.6",
-    "@ensdomains/ui": "1.1.20",
+    "@ensdomains/ui": "1.1.22",
     "apollo-cache-inmemory": "^1.2.9",
     "apollo-client": "^2.4.5",
     "apollo-link": "^1.2.2",
