+++ src/components/SingleName/Address/index.js
@@ -1,0 +1,3 @@
+import Address from './Address'
+
+export default Address
