+++ src/components/SingleName/KeyValueRecord/index.js
@@ -1,0 +1,3 @@
+import KeyValueRecord from './KeyValueRecord'
+
+export default KeyValueRecord
