+++ src/graphql/queries.js
@@ -40,6 +40,12 @@ export const GET_TEXT = gql`
   }
 `
 
+export const GET_ADDR = gql`
+  query getAddr($name: String, $key: String) {
+    getAddr(name: $name, key: $key) @client
+  }
+`
+
 export const GET_SINGLE_NAME = gql`
   query singleName($name: String) {
     singleName(name: $name) @client {
